+++ src/trace-driven/ISA_Def/trace_opcode.h
@@ -4,7 +4,7 @@
 #ifndef TRACE_OPCODE_H
 #define TRACE_OPCODE_H
 
-#include "../abstract_hardware_model.h"
+#include "../../abstract_hardware_model.h"
 #include <unordered_map>
 #include <string>
