--- src/trace-driven/traces-generator/README
@@ -1,1 +1,0 @@
-TO DO
