--- src/trace-driven/traces-generator/traces-post-processing/Makefile
@@ -1,11 +1,0 @@
-TARGET := post-traces-processing
-
-$(TARGET): post-traces-processing.cpp
-	g++ -o $@ $^
-
-run: $(TARGET)
-	./$(TARGET)
-
-clean: 
-	rm -f $(TARGET) *.o
-
